+++ test/inputs.test.ts
@@ -1,3 +1,5 @@
+import * as yaml from "js-yaml";
+
 import * as core from "./mocks/@actions/core.mock";
 
 jest.mock("@actions/core", () => core);
@@ -46,9 +48,33 @@ describe("ActionConfig", () => {
     });
 
     test("construct with an object specifying dry-run", () => {
-      const config: RawActionConfig = { "dry-run": "true" };
+      const config: RawActionConfig = { "dry-run": true };
       expect(() => new ActionConfig(config)).not.toThrow();
     });
+
+  });
+
+  describe(".getSvgoOptionsPath", () => {
+
+    const testPaths = test.each([".svgo.yml", "foo.yml", "in/folder/config.yml"]);
+
+    testPaths("svgo-options is set (to '%s') in the workflow file", (path) => {
+      const instance: ActionConfig = new ActionConfig();
+      core.getInput.mockReturnValueOnce(path);
+
+      const result = instance.getSvgoOptionsPath();
+      expect(result).toBe(path);
+    });
+
+    testPaths("svgo-options is set (to '%s') in the config object", (path) => {
+      const instance: ActionConfig = new ActionConfig({ "svgo-options": path });
+      core.getInput.mockReturnValueOnce(`dir/${path}`);
+
+      const result = instance.getSvgoOptionsPath();
+      expect(result).toBe(path);
+    });
+
+    afterAll(core.getInput.mockReset);
 
   });
 
@@ -86,7 +112,8 @@ describe("ActionConfig", () => {
     });
 
     test("dry-run is `false` in the config object", () => {
-      const instance: ActionConfig = new ActionConfig({ "dry-run": "false" });
+      const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: false");
+      const instance: ActionConfig = new ActionConfig(rawConfig);
       core.getInput.mockReturnValueOnce("true");
 
       const result = instance.isDryRun();
@@ -94,7 +121,26 @@ describe("ActionConfig", () => {
     });
 
     test("dry-run is `true` in the config object", () => {
-      const instance: ActionConfig = new ActionConfig({ "dry-run": "true" });
+      const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: true");
+      const instance: ActionConfig = new ActionConfig(rawConfig);
+      core.getInput.mockReturnValueOnce("false");
+
+      const result = instance.isDryRun();
+      expect(result).toBe(true);
+    });
+
+    test("dry-run is `'false'` in the config object", () => {
+      const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: 'false'");
+      const instance: ActionConfig = new ActionConfig(rawConfig);
+      core.getInput.mockReturnValueOnce("true");
+
+      const result = instance.isDryRun();
+      expect(result).toBe(false);
+    });
+
+    test("dry-run is `'true'` in the config object", () => {
+      const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: 'true'");
+      const instance: ActionConfig = new ActionConfig(rawConfig);
       core.getInput.mockReturnValueOnce("false");
 
       const result = instance.isDryRun();
@@ -102,13 +148,15 @@ describe("ActionConfig", () => {
     });
 
     testNonBoolean("dry run is `'%s'` in the config object", async (value) => {
-      const instance: ActionConfig = new ActionConfig({ "dry-run": value });
+      const rawConfig: RawActionConfig = yaml.safeLoad(`dry-run: '${value}'`);
+      const instance: ActionConfig = new ActionConfig(rawConfig);
       core.getInput.mockReturnValueOnce("false");
 
       const result = instance.isDryRun();
       expect(result).toBe(true);
-      expect(core.info).toHaveBeenCalledTimes(1);
     });
+
+    afterAll(core.getInput.mockReset);
 
   });