+++ CHANGELOG.md
@@ -7,8 +7,15 @@ Versioning].
 
 ## [Unreleased]
 
+- Configure the SVGO options file. ([#99])
+- Fix `required` value for "configuration-path" input. ([#100])
+- Fix interpretation of "dry-run" input in config file. ([#103])
+
+## [0.2.1] - 2020-03-02
+
 - Disable the action from a commit message. ([#91])
 - Add support for a separate configuration file. ([#92])
+- Improve the default commit message. ([#97])
 
 ## [0.2.0] - 2020-03-01
 
@@ -47,3 +54,7 @@ Versioning].
 [#84]: https://github.com/ericcornelissen/svgo-action/pull/84
 [#91]: https://github.com/ericcornelissen/svgo-action/pull/91
 [#92]: https://github.com/ericcornelissen/svgo-action/pull/92
+[#97]: https://github.com/ericcornelissen/svgo-action/pull/97
+[#99]: https://github.com/ericcornelissen/svgo-action/pull/99
+[#100]: https://github.com/ericcornelissen/svgo-action/pull/100
+[#103]: https://github.com/ericcornelissen/svgo-action/pull/103